Separate files for each object type
